refactor: constant size
