+++ src/polodb_core/journal/mod.rs
@@ -1,4 +1,5 @@
 
 mod journal;
+mod frame_header;
 
 pub use journal::{JournalManager, TransactionType};
